ST3: Fix colorpicker binary missing on linux/osx
This patch unifies the behavior of colorpicker binary reloading on both
Sublime Text 2 and 3. From now on the binary will be copied to
Packages/User/ColorPicker/bin/

We first check whether or not the binary exists on plugin loaded. If it
doesn't, or it's older than the package's, then copy
Packages/ColorPicker/lib/<colorpicker> to Packages/User/ColorPicker/bin
This will be the case on both ST2 & 3, if the source binary exists in a
folder-form of package.

On ST3, if only Installed\ Packages/ColorPicker.sublime-packge exists,
and the sublime-package file is newer than the existing binary, then
extract the new one from it.